feat: use `jiti` for typescript config in Node.js environment
